requirements.txt
# Core Django
Django==4.2.16
djangorestframework==3.15.1

# Database
psycopg2-binary==2.9.9

# Redis & Queue
redis==5.0.8
django-rq==2.11.0

# Environment variables
python-decouple==3.8
dj-database-url==2.1.0

# API Documentation
drf-spectacular==0.27.2

